+++ AppObjects/FlatCAMObj.py
@@ -459,8 +459,13 @@ class FlatCAMObj(QtCore.QObject):
     def visible(self, value, threaded=True):
         log.debug("FlatCAMObj.visible()")
 
-        # self.shapes.visible = value   # maybe this is slower in VisPy
-        self.shapes.enabled = value
+        # self.shapes.visible = value   # maybe this is slower in VisPy? use enabled property?
+        if self.shapes.visible is True:
+            if value is False:
+                self.shapes.visible = False
+        else:
+            if value is True:
+                self.shapes.visible = True
 
         if self.app.is_legacy is False:
             # Not all object types has annotations